RUMM-320 Architecture cleanup
