+++ dd-sdk-android-androidx-fragment/src/test/kotlin/com/datadog/android/androidx/fragment/FragmentViewTrackingStrategyTest.kt
@@ -9,8 +9,10 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.fragment.app.FragmentManager
 import com.datadog.android.androidx.fragment.internal.OreoFragmentLifecycleCallbacks
-import com.datadog.android.androidx.fragment.internal.RumMonitorBasedTest
 import com.datadog.android.androidx.fragment.internal.resolveViewName
+import com.datadog.android.androidx.fragment.internal.utils.mockRumMonitor
+import com.datadog.android.androidx.fragment.internal.utils.resetRumMonitorToDefaults
+import com.datadog.android.rum.RumMonitor
 import com.datadog.tools.unit.annotations.TestTargetApi
 import com.datadog.tools.unit.extensions.ApiLevelExtension
 import com.nhaarman.mockitokotlin2.argumentCaptor
@@ -25,6 +27,7 @@ import com.nhaarman.mockitokotlin2.whenever
 import fr.xgouchet.elmyr.Forge
 import fr.xgouchet.elmyr.junit5.ForgeExtension
 import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
@@ -40,7 +43,7 @@ import org.mockito.quality.Strictness
     ExtendWith(ApiLevelExtension::class)
 )
 @MockitoSettings(strictness = Strictness.LENIENT)
-internal class FragmentViewTrackingStrategyTest : RumMonitorBasedTest() {
+internal class FragmentViewTrackingStrategyTest {
     lateinit var underTest: FragmentViewTrackingStrategy
 
     @Mock
@@ -61,16 +64,23 @@ internal class FragmentViewTrackingStrategyTest : RumMonitorBasedTest() {
     @Mock
     lateinit var mockBadContext: Context
 
+    lateinit var mockRumMonitor: RumMonitor
+
     // region Strategy tests
 
     @BeforeEach
-    override fun `set up`(forge: Forge) {
-        super.`set up`(forge)
+    fun `set up`(forge: Forge) {
+        mockRumMonitor = mockRumMonitor()
         whenever(mockAndroidxActivity.supportFragmentManager)
             .thenReturn(mockAndroidxFragmentManager)
         whenever(mockActivity.fragmentManager)
             .thenReturn(mockDefaultFragmentManager)
         underTest = FragmentViewTrackingStrategy(true)
+    }
+
+    @AfterEach
+    fun `tear down`() {
+        resetRumMonitorToDefaults()
     }
 
     @Test
