+++ dd-sdk-android-androidx-fragment/src/test/kotlin/com/datadog/android/androidx/fragment/internal/CompatFragmentLifecycleCallbacksTest.kt
@@ -3,11 +3,15 @@ package com.datadog.android.androidx.fragment.internal
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.fragment.app.FragmentManager
+import com.datadog.android.androidx.fragment.internal.utils.mockRumMonitor
+import com.datadog.android.androidx.fragment.internal.utils.resetRumMonitorToDefaults
+import com.datadog.android.rum.RumMonitor
 import com.nhaarman.mockitokotlin2.eq
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
 import fr.xgouchet.elmyr.Forge
+import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
@@ -21,7 +25,7 @@ import org.mockito.quality.Strictness
     ExtendWith(MockitoExtension::class)
 )
 @MockitoSettings(strictness = Strictness.LENIENT)
-internal class CompatFragmentLifecycleCallbacksTest : RumMonitorBasedTest() {
+internal class CompatFragmentLifecycleCallbacksTest {
     lateinit var underTest: CompatFragmentLifecycleCallbacks
 
     @Mock
@@ -35,12 +39,19 @@ internal class CompatFragmentLifecycleCallbacksTest : RumMonitorBasedTest() {
 
     lateinit var attributesMap: Map<String, Any?>
 
+    lateinit var mockRumMonitor: RumMonitor
+
     @BeforeEach
-    override fun `set up`(forge: Forge) {
-        super.`set up`(forge)
+    fun `set up`(forge: Forge) {
+        mockRumMonitor = mockRumMonitor()
         whenever(mockFragmentActivity.supportFragmentManager).thenReturn(mockFragmentManager)
         attributesMap = forge.aMap { forge.aString() to forge.aString() }
         underTest = CompatFragmentLifecycleCallbacks { attributesMap }
+    }
+
+    @AfterEach
+    fun `tear down`() {
+        resetRumMonitorToDefaults()
     }
 
     @Test
