+++ dd-sdk-android-androidx-fragment/src/test/kotlin/com/datadog/android/androidx/fragment/internal/OreoFragmentLifecycleCallbacksTest.kt
@@ -4,6 +4,8 @@ import android.app.Activity
 import android.app.Fragment
 import android.app.FragmentManager
 import android.os.Build
+import com.datadog.android.androidx.fragment.internal.utils.resetRumMonitorToDefaults
+import com.datadog.android.rum.RumMonitor
 import com.datadog.tools.unit.annotations.TestTargetApi
 import com.datadog.tools.unit.extensions.ApiLevelExtension
 import com.nhaarman.mockitokotlin2.eq
@@ -12,6 +14,7 @@ import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.verifyZeroInteractions
 import com.nhaarman.mockitokotlin2.whenever
 import fr.xgouchet.elmyr.Forge
+import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
@@ -26,7 +29,7 @@ import org.mockito.quality.Strictness
     ExtendWith(ApiLevelExtension::class)
 )
 @MockitoSettings(strictness = Strictness.LENIENT)
-internal class OreoFragmentLifecycleCallbacksTest : RumMonitorBasedTest() {
+internal class OreoFragmentLifecycleCallbacksTest {
     lateinit var underTest: OreoFragmentLifecycleCallbacks
 
     @Mock
@@ -40,12 +43,18 @@ internal class OreoFragmentLifecycleCallbacksTest : RumMonitorBasedTest() {
 
     lateinit var attributesMap: Map<String, Any?>
 
+    lateinit var mockRumMonitor: RumMonitor
+
     @BeforeEach
-    override fun `set up`(forge: Forge) {
-        super.`set up`(forge)
+    fun `set up`(forge: Forge) {
         whenever(mockActivity.fragmentManager).thenReturn(mockFragmentManager)
         attributesMap = forge.aMap { forge.aString() to forge.aString() }
         underTest = OreoFragmentLifecycleCallbacks { attributesMap }
+    }
+
+    @AfterEach
+    fun `tear down`() {
+        resetRumMonitorToDefaults()
     }
 
     @Test
