+++ dd-sdk-android-navigation/src/main/kotlin/com/datadog/android/androidx/navigation/NavigationViewTrackingStrategy.kt
@@ -69,7 +69,7 @@ class NavigationViewTrackingStrategy(
     ) {
         controller.currentDestination?.let { GlobalRum.get().stopView(it) }
 
-        val attributes = if (trackArguments) asRumAttributes(arguments) else emptyMap()
+        val attributes = if (trackArguments) convertToRumAttributes(arguments) else emptyMap()
         val name = destination.getRumViewName()
 
         GlobalRum.get().startView(destination, name, attributes)
