+++ dd-sdk-android-support-fragment/src/test/kotlin/com/datadog/android/support/fragment/internal/OreoFragmentLifecycleCallbacksTest.kt
@@ -4,8 +4,10 @@ import android.app.Activity
 import android.app.Fragment
 import android.app.FragmentManager
 import android.os.Build
+import com.datadog.android.androidx.fragment.internal.utils.mockRumMonitor
+import com.datadog.android.androidx.fragment.internal.utils.resetRumMonitorToDefaults
+import com.datadog.android.rum.RumMonitor
 import com.datadog.android.support.fragment.internal.OreoFragmentLifecycleCallbacks
-import com.datadog.android.support.fragment.internal.RumMonitorBasedTest
 import com.datadog.android.support.fragment.internal.resolveViewName
 import com.datadog.tools.unit.annotations.TestTargetApi
 import com.datadog.tools.unit.extensions.ApiLevelExtension
@@ -15,6 +17,7 @@ import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.verifyZeroInteractions
 import com.nhaarman.mockitokotlin2.whenever
 import fr.xgouchet.elmyr.Forge
+import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
@@ -29,7 +32,7 @@ import org.mockito.quality.Strictness
     ExtendWith(ApiLevelExtension::class)
 )
 @MockitoSettings(strictness = Strictness.LENIENT)
-internal class OreoFragmentLifecycleCallbacksTest : RumMonitorBasedTest() {
+internal class OreoFragmentLifecycleCallbacksTest {
     lateinit var underTest: OreoFragmentLifecycleCallbacks
 
     @Mock
@@ -43,12 +46,19 @@ internal class OreoFragmentLifecycleCallbacksTest : RumMonitorBasedTest() {
 
     lateinit var attributesMap: Map<String, Any?>
 
+    lateinit var mockRumMonitor: RumMonitor
+
     @BeforeEach
-    override fun `set up`(forge: Forge) {
-        super.`set up`(forge)
+    fun `set up`(forge: Forge) {
+        mockRumMonitor = mockRumMonitor()
         whenever(mockActivity.fragmentManager).thenReturn(mockFragmentManager)
         attributesMap = forge.aMap { forge.aString() to forge.aString() }
         underTest = OreoFragmentLifecycleCallbacks { attributesMap }
+    }
+
+    @AfterEach
+    fun `tear down`() {
+        resetRumMonitorToDefaults()
     }
 
     @Test
