+++ dd-sdk-android/src/main/kotlin/com/datadog/android/rum/ActivityLifecycleTrackingStrategy.kt
@@ -72,7 +72,7 @@ abstract class ActivityLifecycleTrackingStrategy : Application.ActivityLifecycle
      * Maps the Bundle key - value properties into compatible attributes for the Rum Events.
      * @param bundle the Bundle we need to transform. Returns an empty Map if this is null.
      */
-    protected fun asRumAttributes(bundle: Bundle?): Map<String, Any?> {
+    protected fun convertToRumAttributes(bundle: Bundle?): Map<String, Any?> {
         if (bundle == null) return emptyMap()
 
         val attributes = mutableMapOf<String, Any?>()
